+++ src/test/groovy/org/gradlex/javaecosystem/capabilities/test/CustomRulesTest.groovy
@@ -199,6 +199,9 @@ compileClasspath - Compile classpath for source set 'main'.
 |    +--- commons-io:commons-io:2.15.0
 |    +--- com.zaxxer:SparseBitSet:1.3
 |    +--- org.apache.logging.log4j:log4j-api:2.21.1
+|    |    \\--- org.apache.logging.log4j:log4j-bom:2.21.1
+|    |         +--- org.apache.logging.log4j:log4j-bom:2.21.1 (*)
+|    |         \\--- org.apache.logging.log4j:log4j-api:2.21.1 (c)
 |    +--- org.apache.poi:poi:5.2.5 (c)
 |    +--- org.apache.poi:poi-excelant:5.2.5 (c)
 |    +--- org.apache.poi:poi-ooxml:5.2.5 (c)
@@ -210,12 +213,12 @@ compileClasspath - Compile classpath for source set 'main'.
 |    |    +--- org.apache.poi:poi:5.2.5 (*)
 |    |    +--- org.apache.poi:poi-ooxml-lite:5.2.5
 |    |    |    \\--- org.apache.xmlbeans:xmlbeans:5.2.0
-|    |    |         \\--- org.apache.logging.log4j:log4j-api:2.21.1
+|    |    |         \\--- org.apache.logging.log4j:log4j-api:2.21.1 (*)
 |    |    +--- org.apache.xmlbeans:xmlbeans:5.2.0 (*)
 |    |    +--- org.apache.commons:commons-compress:1.25.0
 |    |    +--- commons-io:commons-io:2.15.0
 |    |    +--- com.github.virtuald:curvesapi:1.08
-|    |    +--- org.apache.logging.log4j:log4j-api:2.21.1
+|    |    +--- org.apache.logging.log4j:log4j-api:2.21.1 (*)
 |    |    +--- org.apache.commons:commons-collections4:4.4
 |    |    +--- org.apache.poi:poi:5.2.5 (c)
 |    |    +--- org.apache.poi:poi-excelant:5.2.5 (c)
@@ -228,7 +231,7 @@ compileClasspath - Compile classpath for source set 'main'.
 +--- org.apache.poi:poi-ooxml -> 5.2.5 (*)
 \\--- org.apache.poi:poi-scratchpad -> 5.2.5
      +--- org.apache.poi:poi:5.2.5 (*)
-     +--- org.apache.logging.log4j:log4j-api:2.21.1
+     +--- org.apache.logging.log4j:log4j-api:2.21.1 (*)
      +--- org.apache.commons:commons-math3:3.6.1
      +--- commons-codec:commons-codec:1.16.0
      +--- org.apache.poi:poi:5.2.5 (c)